requirements.txt
python-barcode==0.9.0
9==9
asgiref==3.7.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.0.8
django-cors-headers==4.4.0
django-environ==0.11.2
django-tenants==3.6.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
environ==1.0
google-api-core==2.17.1
google-api-python-client==2.122.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
httplib2==0.22.0
idna==3.6
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysqlclient==2.2.4
openapi-codec==1.3.2
pillow==10.4.0
protobuf==4.25.3
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyJWT==2.8.0
pyparsing==3.1.2
pytz==2024.1
requests==2.31.0
rsa==4.9
simplejson==3.19.2
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
